Create a connection to DataSpace

Constructor for DataSpaceConnection-class

CreateConnection(study = NULL, login = NULL, password = NULL,
  verbose = FALSE)

Arguments

study
A character. Name of the study to connect.

login
A character. Optional argument. If there is no netrc file a temporary one can be written by passing login and password of an active DataSpace account.

password
A character. Optional. The password for the selected login.

verbose
A logical. Whether to print the extra details for troubleshooting.

Value

an instance of DataSpaceConnection

Details

Instantiates an DataSpaceConnection for study The constructor will try to take the values of the various `labkey.*` parameters from the global environment. If they don't exist, it will use default values. These are assigned to `options`, which are then used by the DataSpaceConnection class.

See also

DataSpaceConnection-class

Examples

not_run({
  # Single study
  con <- CreateConnection("cvd408")

  # Cross study
  con <- CreateConnection("")
})