The DataSpaceConnection class

The DataSpaceConnection class

The DataSpaceConnection class

Constructor

connectDS

See also

Active bindings

config

A list. Stores configuration of the connection object such as URL, path and username.

availableStudies

A data.table. The table of available studies.

availableGroups

A data.table. The table of available groups.

availablePublications

A data.table. The table of available publications.

mabGridSummary

A data.table. The filtered grid with updated n_ columns and geometric_mean_curve_ic50.

mabGrid

A data.table. The filtered mAb grid.

virusMetadata

A data.table. Metadata about all viruses in the DataSpace.

Methods

Public methods

DataSpaceConnection$new()

DataSpaceConnection$print()

DataSpaceConnection$getStudy()

DataSpaceConnection$getGroup()

DataSpaceConnection$filterMabGrid()

DataSpaceConnection$resetMabGrid()

DataSpaceConnection$getMab()

DataSpaceConnection$downloadPublicationData()

DataSpaceConnection$refresh()

DataSpaceConnection$clone()

Method new()

Initialize a DataSpaceConnection object. See connectDS.

Usage

DataSpaceConnection$new(
  login = NULL,
  password = NULL,
  verbose = FALSE,
  onStaging = FALSE
)

Arguments

login

A character. Optional argument. If there is no netrc file a temporary one can be written by passing login and password of an active DataSpace account.

password

A character. Optional. The password for the selected login.

verbose

A logical. Whether to print the extra details for troubleshooting.

onStaging

A logical. Whether to connect to the staging server instead of the production server.

Returns

A new `DataSpaceConnection` object.

Method print()

Print the DataSpaceConnection object.

Usage

DataSpaceConnection$print()

Method getStudy()

Create a DataSpaceStudy object.

Usage

DataSpaceConnection$getStudy(study)

Arguments

study

A character. Name of the study to retrieve.

Method getGroup()

Create a DataSpaceStudy object.

Usage

DataSpaceConnection$getGroup(groupId)

Arguments

groupId

An integer. ID of the group to retrieve.

Method filterMabGrid()

Filter rows in the mAb grid by specifying the values to keep in the columns found in the mabGrid field. It takes the column and the values and filters the underlying tables.

Usage

DataSpaceConnection$filterMabGrid(using, value)

Arguments

using

A character. Name of the column to filter.

value

A character vector. Values to keep in the mAb grid.

Method resetMabGrid()

Reset the mAb grid to the unfiltered state.

Usage

DataSpaceConnection$resetMabGrid()

Method getMab()

Create a DataSpaceMab object.

Usage

DataSpaceConnection$getMab()

Method downloadPublicationData()

Download publication data for a chosen publication.

Usage

DataSpaceConnection$downloadPublicationData(
  publicationID,
  outputDir = getwd(),
  unzip = TRUE,
  verbose = TRUE
)

Arguments

publicationID

A character/integer. ID for the publication to download data for.

outputDir

A character. Path to directory to download publication data.

unzip

A logical. If TRUE, unzip publication data to outputDir.

verbose

A logical. Default TRUE.

Method refresh()

Refresh the connection object to update available studies and groups.

Usage

DataSpaceConnection$refresh()

Method clone()

The objects of this class are cloneable with this method.

Usage

DataSpaceConnection$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
# Create a connection (Initiate a DataSpaceConnection object)
con <- connectDS()
con

# Connect to cvd408
# https://dataspace.cavd.org/cds/CAVD/app.view#learn/learn/Study/cvd408?q=408
cvd408 <- con$getStudy("cvd408")

# Connect to all studies
cvd <- con$getStudy("cvd408")

# Connect to the NYVAC durability comparison group
# https://dataspace.cavd.org/cds/CAVD/app.view#group/groupsummary/220
nyvac <- con$getGroup(220)

# Refresh the connection object to update available studies and groups
con$refresh()
}