The DataSpaceDonors class

An R6 class for DataSpace MAb Donor data.

Constructor

DataSpaceConnection$getMab()

See also

Super class

DataSpaceConnection -> DataSpaceDonors

Active bindings

mabMetadata

A data.table of mAbs with metadata found in the object.

donorMetadata

A data.table of donors with metadata found in the object.

datasets

A list of data.table objects containing the related data loaded.

variableDefinitions

A data.table of variable definitions.

Methods

Public methods

DataSpaceDonors$new()

DataSpaceDonors$print()

DataSpaceDonors$loadDaash()

DataSpaceDonors$refresh()

DataSpaceDonors$clone()

Inherited methods

DataSpaceConnection$downloadPublicationData()

DataSpaceConnection$filterMabGrid()

DataSpaceConnection$getDaash()

DataSpaceConnection$getDonors()

DataSpaceConnection$getGroup()

DataSpaceConnection$getGroups()

DataSpaceConnection$getMab()

DataSpaceConnection$getMabs()

DataSpaceConnection$getStudies()

DataSpaceConnection$getStudy()

DataSpaceConnection$loadLanlMabMetadata()

DataSpaceConnection$resetMabGrid()

DataSpaceDonors$new()

Initialize DataSpaceMab object. See DataSpaceConnection.

Usage

DataSpaceDonors$new(donorIds)

Arguments

donorIds

a character vector of `donor_id` values.

DataSpaceDonors$print()

Print the DataSpaceMab object summary.

Usage

DataSpaceDonors$print()

DataSpaceDonors$loadDaash()

Load DAASH data to the object.

Usage

DataSpaceDonors$loadDaash()

DataSpaceDonors$refresh()

Refresh the `DataSpaceDonors` object to update datasets.

Usage

DataSpaceDonors$refresh()

DataSpaceDonors$clone()

The objects of this class are cloneable with this method.

Usage

DataSpaceDonors$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) { # \dontrun{
# Create a connection (Initiate a DataSpaceConnection object)
con <- connectDS()

# Print available donors to the console
con$availableDonors

# Query the available donors object and pass that to `getDonors` to get a DataSpaceDonors object
donors <- con$availableDonors[lineage_sequences_available == TRUE & donor_clade == "B",] |>
  con$getDonors()

# Load DAASH data to the object
donors$loadDaash()

} # }