The DataSpaceMab class

The DataSpaceMab class

The DataSpaceMab class

Constructor

DataSpaceConnection$getMab()

See also

Active bindings

config

A list. Stores configuration of the connection object such as URL, path and username.

studyAndMabs

A data.table. The table of available mAbs by study.

mabs

A data.table. The table of available mAbs and their attributes.

nabMab

A data.table. The table of mAbs and their neutralizing measurements against viruses.

studies

A data.table. The table of available studies.

assays

A data.table. The table of assay status by study.

variableDefinitions

A data.table. The table of variable definitions.

Methods

Public methods

DataSpaceMab$new()

DataSpaceMab$print()

DataSpaceMab$refresh()

DataSpaceMab$clone()

Method new()

Initialize DataSpaceMab object. See DataSpaceConnection.

Usage

DataSpaceMab$new(mabMixture, filters, config)

Arguments

mabMixture

A character vector.

filters

A list.

config

A list.

Method print()

Print the DataSpaceMab object summary.

Usage

DataSpaceMab$print()

Method refresh()

Refresh the DataSpaceMab object to update datasets.

Usage

DataSpaceMab$refresh()

Method clone()

The objects of this class are cloneable with this method.

Usage

DataSpaceMab$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
# Create a connection (Initiate a DataSpaceConnection object)
con <- connectDS()

# Browse the mAb Grid
con$mabGridSummary

# Filter the grid by viruses
con$filterMabGrid(using = "virus", value = c("242-14", "Q23.17", "6535.3", "BaL.26", "DJ263.8"))

# Filter the grid by donor species (llama)
con$filterMabGrid(using = "donor_species", value = "llama")

# Check the updated grid
con$mabGridSummary

# Retrieve available viruses in the filtered grid
con$mabGrid[, unique(virus)]

# Retrieve available clades for 1H9 mAb mixture in the filtered grid
con$mabGrid[mab_mixture == "1H9", unique(clade)]

# Create a DataSpaceMab object that contains the filtered mAb data
mab <- con$getMab()
mab

# Inspect the `nabMab` field
mab$nabMab
}