The DataSpaceStudy class

The DataSpaceStudy class

The DataSpaceStudy class

Constructor

DataSpaceConnection$getStudy() DataSpaceConnection$getGroup()

See also

Active bindings

study

A character. The study name.

config

A list. Stores configuration of the connection object such as URL, path and username.

availableDatasets

A data.table. The table of datasets available in the DataSpaceStudy object.

cache

A list. Stores the data to avoid downloading the same tables multiple times.

dataDir

A character. Default directory for storing nonstandard datasets. Set with setDataDir(dataDir).

treatmentArm

A data.table. The table of treatment arm information for the connected study. Not available for all study connection.

group

A character. The group name.

studyInfo

A list. Stores the information about the study.

Methods

Public methods

DataSpaceStudy$new()

DataSpaceStudy$print()

DataSpaceStudy$getDataset()

DataSpaceStudy$clearCache()

DataSpaceStudy$getDatasetDescription()

DataSpaceStudy$setDataDir()

DataSpaceStudy$refresh()

DataSpaceStudy$clone()

Method new()

Initialize DataSpaceStudy class. See DataSpaceConnection.

Usage

DataSpaceStudy$new(study = NULL, config = NULL, group = NULL, studyInfo = NULL)

Arguments

study

A character. Name of the study to retrieve.

config

A list. Stores configuration of the connection object such as URL, path and username.

group

An integer. ID of the group to retrieve.

studyInfo

A list. Stores the information about the study.

Method print()

Print DataSpaceStudy class.

Usage

DataSpaceStudy$print()

Method getDataset()

Get a dataset from the connection.

Usage

DataSpaceStudy$getDataset(
  datasetName,
  mergeExtra = FALSE,
  colFilter = NULL,
  reload = FALSE,
  outputDir = NULL,
  ...
)

Arguments

datasetName

A character. Name of the dataset to retrieve. Accepts the value in either the "name" or "label" field from availableDatasets.

mergeExtra

A logical. If set to TRUE, merge extra information. Ignored for non-integrated datasets.

colFilter

A matrix. A filter as returned by Rlabkey's makeFilter.

reload

A logical. If set to TRUE, download the dataset, whether a cached version exist or not.

outputDir

A character. Optional, specifies directory to download nonstandard datasets. If NULL, data will be downloaded to dataDir, set with setDataDir(dataDir). If dataDir is not set, and outputDir is NULL, a tmp directory will be used.

...

Extra arguments to be passed to labkey.selectRows

Method clearCache()

Clear cache. Remove downloaded datasets.

Usage

DataSpaceStudy$clearCache()

Method getDatasetDescription()

Get variable information.

Usage

DataSpaceStudy$getDatasetDescription(datasetName, outputDir = NULL)

Arguments

datasetName

A character. Name of the dataset to retrieve. Accepts the value in either the "name" or "label" field from availableDatasets.

outputDir

A character. Directory path.

Method setDataDir()

Set default directory to download non-integrated datasets. If no dataDir is set, a tmp directory will be used.

Usage

DataSpaceStudy$setDataDir(dataDir)

Arguments

dataDir

A character. Directory path.

Method refresh()

Refresh the study object to update available datasets and treatment info.

Usage

DataSpaceStudy$refresh()

Method clone()

The objects of this class are cloneable with this method.

Usage

DataSpaceStudy$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
# Create a connection (Initiate a DataSpaceConnection object)
con <- connectDS()

# Connect to cvd408 (Initiate a DataSpaceStudy object)
# https://dataspace.cavd.org/cds/CAVD/app.view#learn/learn/Study/cvd408?q=408
cvd408 <- con$getStudy("cvd408")
cvd408

# Retrieve Neutralizing antibody dataset (NAb) for cvd408 from DataSpace
NAb <- cvd408$getDataset("NAb")

# Get variable information of the NAb dataset
cvd408$getDatasetDescription("NAb")

# Take a look at cvd408's treatment arm information
cvd408$treatmentArm

# Clear cache of a study object
cvd408$clearCache()

# Connect to the NYVAC durability comparison group
# https://dataspace.cavd.org/cds/CAVD/app.view#group/groupsummary/220
nyvac <- con$getGroup(220)

# Connect to all studies
cvd <- con$getStudy("")

# Refresh the study object to update available datasets and treatment info
cvd$refresh()
}