The DataSpaceConnection class

The DataSpaceConnection class

The DataSpaceConnection class

Value

an instance of DataSpaceConnection

Constructor

connectDS

Fields

config

A list. Stores configuration of the connection object such as URL, path and username.

availableStudies

A data.table. The table of available studies.

availableGroups

A data.table. The table of available groups.

mabGrid

A data.table. The filtered mAb grid.

mabGridSummary

A data.table. The filtered grid with updated n_ columns and geometric_mean_curve_ic50.

Methods

initialize(login = NULL, password = NULL, verbose = FALSE, onStaging = FALSE)

Initialize a DataSpaceConnection object. See connectDS.

print()

Print the DataSpaceConnection object.

getStudy(study, groupId = NULL)

Create a DataSpaceStudy object. study: A character. Name of the study to retrieve.

getGroup(groupId)

Create a DataSpaceStudy object. groupId: An integer. ID of the group to retrieve.

refresh()

Refresh the connection object to update available studies and groups.

filterMabGrid(using, value)

Filter rows in the mAb grid by specifying the values to keep in the columns found in the mabGrid field. It takes the column and the values and filters the underlying tables. using: A character. Name of the column to filter. value: A character vector. Values to keep in the mAb grid.

getMab()

Create a DataSpaceMab object.

resetMabGrid()

Reset the mAb grid to the unfiltered state.

See also

Methods

Public methods

DataSpaceConnection$new()

DataSpaceConnection$print()

DataSpaceConnection$getStudy()

DataSpaceConnection$getGroup()

DataSpaceConnection$filterMabGrid()

DataSpaceConnection$resetMabGrid()

DataSpaceConnection$getMab()

DataSpaceConnection$refresh()

DataSpaceConnection$clone()

Method new()

Usage

DataSpaceConnection$new(
  login = NULL,
  password = NULL,
  verbose = FALSE,
  onStaging = FALSE
)

Method print()

Usage

DataSpaceConnection$print()

Method getStudy()

Usage

DataSpaceConnection$getStudy(study)

Method getGroup()

Usage

DataSpaceConnection$getGroup(groupId)

Method filterMabGrid()

Usage

DataSpaceConnection$filterMabGrid(using, value)

Method resetMabGrid()

Usage

DataSpaceConnection$resetMabGrid()

Method getMab()

Usage

DataSpaceConnection$getMab()

Method refresh()

Usage

DataSpaceConnection$refresh()

Method clone()

The objects of this class are cloneable with this method.

Usage

DataSpaceConnection$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) {
# Create a connection (Initiate a DataSpaceConnection object)
con <- connectDS()
con

# Connect to cvd408
# https://dataspace.cavd.org/cds/CAVD/app.view#learn/learn/Study/cvd408?q=408
cvd408 <- con$getStudy("cvd408")

# Connect to all studies
cvd <- con$getStudy("cvd408")

# Connect to the NYVAC durability comparison group
# https://dataspace.cavd.org/cds/CAVD/app.view#group/groupsummary/220
nyvac <- con$getGroup(220)

# Refresh the connection object to update available studies and groups
con$refresh()
}