The DataSpaceConnection class

The DataSpaceConnection class

DataSpaceConnection

Format

An object of class R6ClassGenerator of length 24.

Details

Uses global variables labkey.url.base, and labkey.url.path, to access a study. labkey.url.base should be https://dataspace.cavd.org/. labkey.url.path should be /CAVD/studyname, where studyname is the accession number of the study. DataSpaceConnection will initialize itself, and look for a .netrc file in "~/" the user's home directory. The .netrc file should contain a machine, login, and password entry to allow access to DataSpace, where machine is the host name like "dataspace.cavd.org".

Constructor

connectDS

Fields

study

A character. The study name. Use an empty string ("") to create a connection at the project level.

config

A list. Stores configuration of the connection object such as URL, path and username.

availableDatasets

A data.frame. The table of datasets available in the connection object.

cache

A list. Stores the data to avoid downloading the same tables multiple times.

treatmentArm

A data.frame. The table of treatment arm information for the connected study. Not available for cross study connection.

Methods

initialize(study = NULL, login = NULL, password = NULL, verbose = FALSE, onStaging = FALSE)

Initialize DataSpaceConnection class. See connectDS.

print()

Print DataSpaceConnection class.

getDataset(datasetName, colFilter = NULL, reload = FALSE, ...)

Get a dataset from the connection. datasetName: A character. Name of the dataset to retrieve. colFilter: A matrix. A filter as returned by Rlabkey's makeFilter. reload: A logical. If set to TRUE, download the dataset, whether a cached version exist or not. ...: Extra arguments to be passed to labkey.selectRows

clearCache()

Clear cache. Remove downloaded datasets.

getVariableInfo(datasetName)

Get variable information. datasetName: A character. Name of the dataset to retrieve.

See also

connectDS DataSpaceR-package

Examples

# NOT RUN {
# Create a connection (Initiate a DataSpaceConnection object)
con <- connectDS("cvd408")

# Retrieve Neutralizing antibody dataset (NAb) from DataSpace
con$getDataset("NAb")
# }