The DataSpaceGroups class

An R6 class for DataSpace Groups data.

Constructor

DataSpaceConnection$getGroups()

See also

Super class

DataSpaceConnection -> DataSpaceGroups

Active bindings

availableDatasets

A data.table of datasets available in the object.

datasets

A list of data.table objects containing the availableDatasets that were loaded.

variableDefinitions

A data.table of variable definitions.

Methods

Public methods

DataSpaceGroups$new()

DataSpaceGroups$print()

DataSpaceGroups$refresh()

DataSpaceGroups$clone()

Inherited methods

DataSpaceConnection$downloadPublicationData()

DataSpaceConnection$filterMabGrid()

DataSpaceConnection$getDaash()

DataSpaceConnection$getDonors()

DataSpaceConnection$getGroup()

DataSpaceConnection$getGroups()

DataSpaceConnection$getMab()

DataSpaceConnection$getMabs()

DataSpaceConnection$getStudies()

DataSpaceConnection$getStudy()

DataSpaceConnection$loadLanlMabMetadata()

DataSpaceConnection$resetMabGrid()

DataSpaceGroups$new()

Initialize `DataSpaceGroups` class. See DataSpaceConnection.

Usage

DataSpaceGroups$new(groupIds = NULL)

Arguments

groupIds

A character vecotor of `group_id` values. as URL, path and username.

DataSpaceGroups$print()

Print DataSpaceStudy class.

Usage

DataSpaceGroups$print()

DataSpaceGroups$refresh()

Refresh loaded integrated datasets, and information of what datasets are available.

Usage

DataSpaceGroups$refresh()

DataSpaceGroups$clone()

The objects of this class are cloneable with this method.

Usage

DataSpaceGroups$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (FALSE) { # \dontrun{
# Create a connection (Initiate a DataSpaceConnection object)
con <- connectDS()

# Get group by `group_id` or pass a filtered `availableGroups` object.
groups <- con$getGroups(c(266, 267))
groups <- con$availableGroups[label == "NYVAC durability comparison"] |>
  con$getGroups()

# Retrieving group assay data for cvd408 from
# DataSpace is done automatically when the groups object is created.
groups$datasets$BAMA

# Get variable information of the assay dataset
groups$datasetDescription$BAMA

} # }